# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillSwap Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (browse for skill matching)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/matches:
    get:
      operationId: getSkillMatches
      tags: [users]
      summary: Get users who match current user's wanted skills
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Matched users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithMatch"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: Get skill swap requests for current user
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sent, received, all]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, completed]
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SwapRequest"

    post:
      operationId: createRequest
      tags: [requests]
      summary: Send a skill swap request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
        "409":
          description: Duplicate request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}:
    get:
      operationId: getRequestById
      tags: [requests]
      summary: Get a request by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /requests/{id}/accept:
    post:
      operationId: acceptRequest
      tags: [requests]
      summary: Accept a skill swap request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
  /requests/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [requests]
      summary: Reject a skill swap request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
  /requests/{id}/complete:
    post:
      operationId: completeRequest
      tags: [requests]
      summary: Mark a request as completed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        bio:
          type: string
        location:
          type: string
        profileImage:
          type: string
        skillsOffered:
          type: array
          items:
            type: string
        skillsWanted:
          type: array
          items:
            type: string
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - skillsOffered
        - skillsWanted
        - isBlocked
        - createdAt
    UserWithMatch:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            matchingSkills:
              type: array
              items:
                type: string
            matchScore:
              type: integer
          required:
            - matchingSkills
            - matchScore
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        location:
          type: string
        profileImage:
          type: string
        skillsOffered:
          type: array
          items:
            type: string
        skillsWanted:
          type: array
          items:
            type: string
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages
...
    SwapRequest:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        skillOffered:
          type: string
        skillRequested:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, completed]
        sender:
          $ref: "#/components/schemas/User"
        receiver:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - senderId
        - receiverId
        - skillOffered
        - skillRequested
        - status
        - createdAt
    CreateRequestInput:
      type: object
      properties:
        receiverId:
          type: integer
        skillOffered:
          type: string
        skillRequested:
          type: string
        message:
          type: string
      required:
        - receiverId
        - skillOffered
        - skillRequested
...