# lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillSwap Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  name?: string;
  bio?: string;
  location?: string;
  profileImage?: string;
  skillsOffered?: string[];
  skillsWanted?: string[];
}