# lib/api-zod/src/generated/types/userListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillSwap Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}